requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
bcrypt==4.3.0
blinker==1.9.0
boto3==1.39.12
botocore==1.39.12
certifi==2025.7.14
click==8.1.8
Deprecated==1.2.18
exceptiongroup==1.3.0
flasgger==0.9.7.1
flask==3.1.1
flask-limiter==3.11.0
h11==0.16.0
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
numpy==2.0.2
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rich==13.9.4
rpds-py==0.26.0
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zipp==3.23.0
